s4:kdc: Use portable format specifier

+++ source4/kdc/kdc-heimdal.c
@@ -81,8 +81,8 @@ static kdc_code kdc_process(struct kdc_server *kdc,
 		return KDC_ERROR;
 	}
 
-	DBG_DEBUG("Received KDC packet of length %lu from %s\n",
-		  (long)input->length - 4, pa);
+	DBG_DEBUG("Received KDC packet of length %zu from %s\n",
+		  input->length - 4, pa);
 
 	ret = krb5_kdc_process_krb5_request(kdc->smb_krb5_context->krb5_context,
 					    kdc_config,
